backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pyjwt==2.9.0
httpx==0.27.2
sqlalchemy[asyncio]==2.0.34
psycopg[binary]==3.1.19
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
boto3==1.35.23
